refactor: proto files for scoped service reflection
