+++ libs/common/src/types/index.ts
@@ -1,7 +1,6 @@
 export * from './proto/auth';
 export * from './proto/chama';
 export * from './proto/lib';
-export * from './proto/lightning';
 export * from './proto/swap';
 export * from './proto/nostr';
 export * from './proto/sms';
