--- proto/lightning.proto
@@ -1,15 +1,0 @@
-syntax = "proto3";
-
-package lightning;
-
-message Bolt11 {
-  // Bolt11 lightning invoice
-  string invoice = 1;
-}
-
-message FmInvoice {
-  // Fallback invoice
-  string invoice = 1;
-
-  string operationId = 2;
-}
